+++ src/Monad.ts
@@ -345,3 +345,53 @@ export function nonePassM<M>(
       ),
     )
 }
+
+/**
+ * Like applicative `when`, but the condition is monadic.
+ *
+ * @example
+ * import { pipe } from 'fp-ts/function'
+ * import { whenM } from 'fp-ts-std/Monad'
+ * import * as IO from 'fp-ts/IO'
+ * import * as IOE from 'fp-ts/IOEither'
+ * import { log } from 'fp-ts/Console'
+ *
+ * const isInvalid = (n: number): IO.IO<boolean> => () => Date.now() !== 42
+ *
+ * pipe(
+ *   IOE.of(123),
+ *   IOE.chainFirstIOK(n =>
+ *     whenM(IO.Monad)(isInvalid(n))(log(n))),
+ * )
+ *
+ * @since 0.16.0
+ */
+export function whenM<F extends URIS4>(
+  M: Monad4<F>,
+): <S, R, E>(
+  b: Kind4<F, S, R, E, boolean>,
+) => (x: Kind4<F, S, R, E, void>) => Kind4<F, S, R, E, void>
+export function whenM<F extends URIS3>(
+  M: Monad3<F>,
+): <R, E>(
+  b: Kind3<F, R, E, boolean>,
+) => (x: Kind3<F, R, E, void>) => Kind3<F, R, E, void>
+export function whenM<F extends URIS3, E>(
+  M: Monad3C<F, E>,
+): <R>(
+  b: Kind3<F, R, E, boolean>,
+) => (x: Kind3<F, R, E, void>) => Kind3<F, R, E, void>
+export function whenM<F extends URIS2>(
+  M: Monad2<F>,
+): <E>(b: Kind2<F, E, boolean>) => (x: Kind2<F, E, void>) => Kind2<F, E, void>
+export function whenM<F extends URIS2, E>(
+  M: Monad2C<F, E>,
+): (b: Kind2<F, E, boolean>) => (x: Kind2<F, E, void>) => Kind2<F, E, void>
+export function whenM<F extends URIS>(
+  M: Monad1<F>,
+): (b: Kind<F, boolean>) => (x: Kind<F, void>) => Kind<F, void>
+export function whenM<F>(
+  M: Monad<F>,
+): (b: HKT<F, boolean>) => (x: HKT<F, void>) => HKT<F, void> {
+  return b => x => M.chain(b, bb => (bb ? x : M.of(undefined)))
+}